requirements.txt
numpy
scipy
sounddevice
soundfile
torch
torchaudio
openai-whisper
piper-tts
